chore: move web utils from web, utils to web-utils
