+++ libtransmission/completion.h
@@ -14,7 +14,6 @@
 
 #include "transmission.h"
 #include "bitfield.h"
-#include "utils.h" /* tr_getRatio() */
 
 struct tr_completion
 {
