+++ libtransmission/metainfo.cc
@@ -25,6 +25,7 @@
 #include "tr-assert.h"
 #include "utils.h"
 #include "variant.h"
+#include "web-utils.h"
 
 using namespace std::literals;
