+++ libtransmission/rpc-server.cc
@@ -20,21 +20,23 @@
 #include <event2/http_struct.h> /* TODO: eventually remove this */
 
 #include "transmission.h"
-#include "crypto.h" /* tr_ssha1_matches() */
+
 #include "crypto-utils.h" /* tr_rand_buffer() */
+#include "crypto.h" /* tr_ssha1_matches() */
 #include "error.h"
 #include "fdlimit.h"
 #include "log.h"
 #include "net.h"
 #include "platform.h" /* tr_getWebClientDir() */
-#include "rpcimpl.h"
 #include "rpc-server.h"
-#include "session.h"
+#include "rpcimpl.h"
 #include "session-id.h"
+#include "session.h"
 #include "tr-assert.h"
 #include "trevent.h"
 #include "utils.h"
 #include "variant.h"
+#include "web-utils.h"
 #include "web.h"
 
 /* session-id is used to make cross-site request forgery attacks difficult.
