+++ libtransmission/rpcimpl.h
@@ -12,11 +12,12 @@
 
 #include "transmission.h"
 #include "tr-macros.h"
-#include "variant.h"
 
 /***
 ****  RPC processing
 ***/
+
+struct tr_variant;
 
 using tr_rpc_response_func = void (*)(tr_session* session, tr_variant* response, void* user_data);
