+++ libtransmission/utils.h
@@ -193,12 +193,12 @@ void tr_free_ptrv(void* const* p);
  */
 void* tr_memdup(void const* src, size_t byteCount);
 
-#define tr_new(struct_type, n_structs) (static_cast<(struct_type*)>(tr_malloc(sizeof(struct_type) * (size_t)(n_structs))))
+#define tr_new(struct_type, n_structs) (static_cast<struct_type*>(tr_malloc(sizeof(struct_type) * (size_t)(n_structs))))
 
-#define tr_new0(struct_type, n_structs) (static_cast<(struct_type)*>(tr_malloc0(sizeof(struct_type) * (size_t)(n_structs))))
+#define tr_new0(struct_type, n_structs) (static_cast<struct_type*>(tr_malloc0(sizeof(struct_type) * (size_t)(n_structs))))
 
 #define tr_renew(struct_type, mem, n_structs) \
-    (static_cast<(struct_type)*>(tr_realloc((mem), sizeof(struct_type) * (size_t)(n_structs))))
+    (static_cast<struct_type*>(tr_realloc((mem), sizeof(struct_type) * (size_t)(n_structs))))
 
 /**
  * @brief make a newly-allocated copy of a substring
@@ -270,36 +270,6 @@ char* tr_strsep(char** str, char const* delim);
 
 void tr_binary_to_hex(void const* input, void* output, size_t byte_length) TR_GNUC_NONNULL(1, 2);
 void tr_hex_to_binary(void const* input, void* output, size_t byte_length) TR_GNUC_NONNULL(1, 2);
-
-/** @brief convenience function to determine if an address is an IP address (IPv4 or IPv6) */
-bool tr_addressIsIP(char const* address);
-
-/** @brief return true if the url is a http or https or UDP url that Transmission understands */
-bool tr_urlIsValidTracker(std::string_view url);
-
-/** @brief return true if the url is a [ http, https, ftp, sftp ] url that Transmission understands */
-bool tr_urlIsValid(std::string_view url);
-
-// TODO: move this to types.h
-struct tr_parsed_url_t
-{
-    std::string_view scheme;
-    std::string_view host;
-    std::string_view path;
-    std::string_view portstr;
-    int port = -1;
-};
-
-std::optional<tr_parsed_url_t> tr_urlParse(std::string_view url);
-
-// like tr_urlParse(), but with the added constraint that 'scheme'
-// must be one we that Transmission supports for announce and scrape
-std::optional<tr_parsed_url_t> tr_urlParseTracker(std::string_view url);
-
-/** @brief parse a URL into its component parts
-    @return True on success or false if an error occurred */
-bool tr_urlParse(char const* url, size_t url_len, char** setme_scheme, char** setme_host, int* setme_port, char** setme_path)
-    TR_GNUC_NONNULL(1);
 
 /** @brief return TR_RATIO_NA, TR_RATIO_INF, or a number in [0..1]
     @return TR_RATIO_NA, TR_RATIO_INF, or a number in [0..1] */
