+++ libtransmission/web.cc
@@ -602,199 +602,3 @@ char const* tr_webGetTaskRealUrl(struct tr_web_task* task)
     curl_easy_getinfo(task->curl_easy, CURLINFO_EFFECTIVE_URL, &url);
     return url;
 }
-
-/*****
-******
-******
-*****/
-
-char const* tr_webGetResponseStr(long code)
-{
-    switch (code)
-    {
-    case 0:
-        return "No Response";
-
-    case 101:
-        return "Switching Protocols";
-
-    case 200:
-        return "OK";
-
-    case 201:
-        return "Created";
-
-    case 202:
-        return "Accepted";
-
-    case 203:
-        return "Non-Authoritative Information";
-
-    case 204:
-        return "No Content";
-
-    case 205:
-        return "Reset Content";
-
-    case 206:
-        return "Partial Content";
-
-    case 300:
-        return "Multiple Choices";
-
-    case 301:
-        return "Moved Permanently";
-
-    case 302:
-        return "Found";
-
-    case 303:
-        return "See Other";
-
-    case 304:
-        return "Not Modified";
-
-    case 305:
-        return "Use Proxy";
-
-    case 306:
-        return " (Unused)";
-
-    case 307:
-        return "Temporary Redirect";
-
-    case 400:
-        return "Bad Request";
-
-    case 401:
-        return "Unauthorized";
-
-    case 402:
-        return "Payment Required";
-
-    case 403:
-        return "Forbidden";
-
-    case 404:
-        return "Not Found";
-
-    case 405:
-        return "Method Not Allowed";
-
-    case 406:
-        return "Not Acceptable";
-
-    case 407:
-        return "Proxy Authentication Required";
-
-    case 408:
-        return "Request Timeout";
-
-    case 409:
-        return "Conflict";
-
-    case 410:
-        return "Gone";
-
-    case 411:
-        return "Length Required";
-
-    case 412:
-        return "Precondition Failed";
-
-    case 413:
-        return "Request Entity Too Large";
-
-    case 414:
-        return "Request-URI Too Long";
-
-    case 415:
-        return "Unsupported Media Type";
-
-    case 416:
-        return "Requested Range Not Satisfiable";
-
-    case 417:
-        return "Expectation Failed";
-
-    case 421:
-        return "Misdirected Request";
-
-    case 500:
-        return "Internal Server Error";
-
-    case 501:
-        return "Not Implemented";
-
-    case 502:
-        return "Bad Gateway";
-
-    case 503:
-        return "Service Unavailable";
-
-    case 504:
-        return "Gateway Timeout";
-
-    case 505:
-        return "HTTP Version Not Supported";
-
-    default:
-        return "Unknown Error";
-    }
-}
-
-void tr_http_escape(struct evbuffer* out, std::string_view str, bool escape_reserved)
-{
-    auto constexpr ReservedChars = std::string_view{ "!*'();:@&=+$,/?%#[]" };
-    auto constexpr UnescapedChars = std::string_view{ "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ-_.~" };
-
-    for (auto& ch : str)
-    {
-        if ((UnescapedChars.find(ch) != std::string_view::npos) || (ReservedChars.find(ch) && !escape_reserved))
-        {
-            evbuffer_add_printf(out, "%c", ch);
-        }
-        else
-        {
-            evbuffer_add_printf(out, "%%%02X", (unsigned)(ch & 0xFF));
-        }
-    }
-}
-
-char* tr_http_unescape(char const* str, size_t len)
-{
-    char* tmp = curl_unescape(str, len);
-    char* ret = tr_strdup(tmp);
-    curl_free(tmp);
-    return ret;
-}
-
-static bool is_rfc2396_alnum(uint8_t ch)
-{
-    return ('0' <= ch && ch <= '9') || ('A' <= ch && ch <= 'Z') || ('a' <= ch && ch <= 'z') || ch == '.' || ch == '-' ||
-        ch == '_' || ch == '~';
-}
-
-void tr_http_escape_sha1(char* out, tr_sha1_digest_t const& digest)
-{
-    for (auto const b : digest)
-    {
-        if (is_rfc2396_alnum(uint8_t(b)))
-        {
-            *out++ = (char)b;
-        }
-        else
-        {
-            out += tr_snprintf(out, 4, "%%%02x", (unsigned int)b);
-        }
-    }
-
-    *out = '\0';
-}
-
-void tr_http_escape_sha1(char* out, uint8_t const* sha1_digest)
-{
-    auto digest = tr_sha1_digest_t{};
-    std::copy_n(reinterpret_cast<std::byte const*>(sha1_digest), std::size(digest), std::begin(digest));
-    tr_http_escape_sha1(out, digest);
-}
